removed anicheck, partitioned the task page, added taskstore. Delete not working
