+++ app/components/task/taskStore.ts
@@ -1,0 +1,23 @@
+import { makeObservable, observable, action, runInAction } from 'mobx';
+
+export interface listItem {
+  title: string;
+  tasks: string[];
+}
+
+class taskStore {
+  list: listItem[] = [];
+
+  constructor() {
+    makeObservable(this, {
+      list: observable,
+      setList: action,
+    });
+  }
+
+  setList(newList: listItem[]) {
+    this.list = newList;
+  }
+}
+
+export default new taskStore();
