+++ package.json
@@ -15,6 +15,8 @@
     "clsx": "^2.1.0",
     "framer-motion": "^11.0.3",
     "lucide-react": "^0.319.0",
+    "mobx": "^6.12.0",
+    "mobx-react": "^9.1.0",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
